+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.11.2] - 2020-02-06
 - Add Monday through Sunday to the X-axis using only the first letter
+- Remove questions without answers on Data Tab
+- Change question width into 100%
 
 ## [0.11.0] - 2020-02-05
 - :arrow_up: upgrade react native version @0.61.5